pyproject.toml
[project]
name = "credit-scoring-tests"
version = "0.1.0"
description = "Autotests for credit scoring module"
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.0.0",
    "allure-pytest>=2.13.0",
    "psycopg2-binary>=2.9.9",
    "pandas>=2.2.0"
]
